More or less works. Archive works after second click.
