+++ src/bookmarksList.html
@@ -25,7 +25,7 @@
       <p>
         <a id="my-items" ng-click="loadUnreadBookmarks()">My items</a> | <!-- TODO Hope that is the right callback -->
         <a id="archived-items" ng-click="loadArchivedBookmarks()">Archived</a> |
-        <a id="favorited-items"> Favorited </a>
+        <a id="favorited-items" ng-click="loadFavoritedBookmarks()"> Favorited </a>
       </p>
